Fix #61
CSL does not distinguish between different kinds of theses. According to the CSL spec 1.0.1 the kind of a thesis shall be specified in the "genre" field. In the case of @mastersthesis records, genre is populated with "Master's thesis", for @phdthesis genre is set to "PhD thesis". For both kinds of thesis, the "genre" field can be overridden by using the type file in BibTeX.