add support for type and key fields
We keep using our own type and key methods but make sure
that importing BibTeX key or type fields works. When
exporting to other key-value formats we use bibtex_key
and bibtex_type for our internal values so as not to
clash with the names used by BibTeX. This breaks backwards
compatibility!

Fix #54