Support Chinese, Japanese and Korean names.
In these language, names are written in <lastname> <firstname>.
BibTeX files written in such a manner are popular in Japan.
Because ruby-bibtex's name lexer first creates the array of tokens prior to parsing,
it is possible to put the letter type information of a name into this array, in the position just before the name.
Racc's grammar is also modified, to switch its behavior by this saved information.
In this implementation, the `AND` token is exploited to store the information.

I couldn't support (and perhaps no software supports) other languages,
e.g. Vietnamese or Hungarian, because detecting those language is not simple.
In future, maybe biblatex's metadata can be used for those.